make notelist flexible
