+++ frontend/src/pages/folder/[id].tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const FolderPage = () => {
+  return <div>FolderPage</div>;
+};
+
+export default FolderPage;
