+++ frontend/src/pages/index.tsx
@@ -1,10 +1,20 @@
 import NoteList from "@/components/NoteList";
-import React from "react";
+import { useNote } from "@/hooks/useNote";
+import { noteListAtom, searchQueryAtom } from "@/jotai/noteAtom";
+import { useAtomValue } from "jotai";
+import React, { useEffect } from "react";
 
 const Home = () => {
+  const notes = useAtomValue(noteListAtom);
+  const searchQuery = useAtomValue(searchQueryAtom);
+  const { setNoteListBySearch } = useNote();
+  useEffect(() => {
+    setNoteListBySearch();
+  }, [searchQuery]);
+
   return (
     <div className="h-screen">
-      <NoteList />
+      <NoteList notes={notes} />
     </div>
   );
 };
